consolidate more notebook stuff into consistet nt pattern
