+++ pgml-dashboard/app/static/js/controllers/notebook-cell.js
