+++ pgml-dashboard/app/static/js/controllers/notebook.js
