+++ pgml-dashboard/app/templates/notebooks/notebooks.html
@@ -1,0 +1,40 @@
+{% extends "base.html" %}
+
+{% block main %}
+<section>
+   <h1>
+    <span class="material-symbols-outlined">newspaper</span>Notebooks
+  </h1> 
+
+  {% if notebooks %}
+  <ol class="object_list">
+    {% for notebook in notebooks %}
+    <li>
+      <a href="{% url 'notebooks/notebook' notebook.pk %}">
+        <span>
+          {{ notebook.name }}
+        </span>
+      </a>
+    </li>
+    {% endfor %}
+  </ol>
+  {% endif %}
+</section>
+
+
+<section>
+  <h2>New notebook</h2>
+
+  <form action="{% url 'notebooks/create' %}" method="post">
+    {% csrf_token %}
+
+    <div class="input-container">
+      <input type="text" name="name" required placeholder="New unique notebook name" size="50" />
+    </div>
+
+    <div class="button-container">
+      <button class="next" type="submit">Create notebook</button>
+    </div>
+  </form>
+</section>
+{% endblock %}
