+++ pgml-dashboard/app/templates/notebooks/sql_error.html
