+++ pgml-dashboard/app/urls.py
@@ -2,18 +2,7 @@ from django.urls import path
 from rest_framework import routers
 
 from app.views import root, projects, models, snapshots, deployments, console, notebooks
-from notebooks.views import (
-    notebook,
-    add_notebook_cell,
-    remove_notebook_cell,
-    notebook_cell,
-    edit_notebook_cell,
-    undo_remove_notebook_cell,
-    create_notebook,
-    rename_notebook,
-    reset_notebook,
-    play_notebook_cell,
-)
+
 
 router = routers.DefaultRouter()
 router.register("projects", projects.ProjectViewSet)
@@ -42,18 +31,18 @@ urlpatterns = [
     path("set-auth-cookie/", root.set_auth_cookie, name="set-auth-cookie"),
 
     path("notebooks/", notebooks.index, name="notebooks"),
-    path("notebooks/<int:pk>/", notebook, name="notebooks/notebook"),
-    path("notebooks/create/", create_notebook, name="notebooks/create"),
-    path("notebooks/<int:pk>/rename/", rename_notebook, name="notebooks/rename"),
-    path("notebooks/<int:pk>/cell/add/", add_notebook_cell, name="notebooks/cell/add"),
-    path("notebooks/<int:notebook_pk>/cell/<int:cell_pk>/", notebook_cell, name="notebooks/cell/get"),
-    path("notebooks/<int:notebook_pk>/cell/<int:cell_pk>/edit/", edit_notebook_cell, name="notebooks/cell/edit"),
-    path("notebooks/<int:notebook_pk>/cell/<int:cell_pk>/remove/", remove_notebook_cell, name="notebooks/cell/remove"),
+    path("notebooks/<int:pk>/", notebooks.notebook, name="notebooks/notebook"),
+    path("notebooks/create/", notebooks.create_notebook, name="notebooks/create"),
+    path("notebooks/<int:pk>/rename/", notebooks.rename_notebook, name="notebooks/rename"),
+    path("notebooks/<int:pk>/cell/add/", notebooks.add_notebook_cell, name="notebooks/cell/add"),
+    path("notebooks/<int:notebook_pk>/cell/<int:cell_pk>/", notebooks.notebook_cell, name="notebooks/cell/get"),
+    path("notebooks/<int:notebook_pk>/cell/<int:cell_pk>/edit/", notebooks.edit_notebook_cell, name="notebooks/cell/edit"),
+    path("notebooks/<int:notebook_pk>/cell/<int:cell_pk>/remove/", notebooks.remove_notebook_cell, name="notebooks/cell/remove"),
     path(
         "notebooks/<int:notebook_pk>/cell/<int:cell_pk>/remove/undo/",
-        undo_remove_notebook_cell,
+        notebooks.undo_remove_notebook_cell,
         name="notebooks/cell/remove/undo",
     ),
-    path("notebooks/<int:pk>/reset/", reset_notebook, name="notebooks/reset"),
-    path("notebooks/<int:notebook_pk>/cell/<int:cell_pk>/play/", play_notebook_cell, name="notebooks/cell/play"),
+    path("notebooks/<int:pk>/reset/", notebooks.reset_notebook, name="notebooks/reset"),
+    path("notebooks/<int:notebook_pk>/cell/<int:cell_pk>/play/", notebooks.play_notebook_cell, name="notebooks/cell/play"),
 ]
