--- pgml-dashboard/notebooks/urls.py
@@ -1,5 +1,0 @@
-from django.urls import path
-
-
-urlpatterns = [
-]
